Changes in callbacks
Changes in getch get and post callbacks and some files directory change
for convenience
