+++ src/auth/src/main/resources/static/js/utils/fetch/data/details/fetch-headers.js
